# WIP: Improved latency of refactored iterators.