--- a/storage/merge.go
+++ b/storage/merge.go
@@ -416,10 +416,7 @@ func (h *genericSeriesSetHeap) Pop() interface{} {
 //
 // This works the best with replicated series, where data from two series are exactly the same. This does not work well
 // with "almost" the same data, e.g. from 2 Prometheus HA replicas. This is fine, since from the Prometheus perspective
-// this never happens.
-//
-// NOTE: Use this merge function only when you see potentially overlapping series, as this introduces a small overhead
-// to handle overlaps between series.
+// this never happens. ChainedSeriesMerge is optimized for non-overlap cases as well.
 func ChainedSeriesMerge(series ...Series) Series {
 	if len(series) == 0 {
 		return nil
@@ -438,10 +435,12 @@ func ChainedSeriesMerge(series ...Series) Series {
 
 // chainSampleIterator is responsible to iterate over samples from different iterators of the same time series in timestamps
 // order. If one or more samples overlap, one sample from random overlapped ones is kept and all others with the same
-// timestamp are dropped.
+// timestamp are dropped. It's optimized for non-overlap cases as well.
 type chainSampleIterator struct {
 	iterators []chunkenc.Iterator
 	h         samplesIteratorHeap
+
+	curr chunkenc.Iterator
 }
 
 func newChainSampleIterator(iterators []chunkenc.Iterator) chunkenc.Iterator {
@@ -462,11 +461,10 @@ func (c *chainSampleIterator) Seek(t int64) bool {
 }
 
 func (c *chainSampleIterator) At() (t int64, v float64) {
-	if len(c.h) == 0 {
-		panic("chainSampleIterator.At() called after .Next() returned false.")
+	if c.curr == nil {
+		panic("chainSampleIterator.At() called before first .Next() or after .Next() returned false.")
 	}
-
-	return c.h[0].At()
+	return c.curr.At()
 }
 
 func (c *chainSampleIterator) Next() bool {
@@ -476,29 +474,43 @@ func (c *chainSampleIterator) Next() bool {
 				heap.Push(&c.h, iter)
 			}
 		}
-
-		return len(c.h) > 0
-	}
-
-	if len(c.h) == 0 {
+		if len(c.h) > 0 {
+			c.curr = heap.Pop(&c.h).(chunkenc.Iterator)
+			return true
+		}
 		return false
 	}
 
-	currt, _ := c.At()
-	for len(c.h) > 0 {
-		nextt, _ := c.h[0].At()
-		// All but one of the overlapping samples will be dropped.
-		if nextt != currt {
-			break
+	for {
+		if c.curr.Next() {
+			if len(c.h) == 0 {
+				return true
+			}
+
+			currt, _ := c.curr.At()
+			nextt, _ := c.h[0].At()
+			if currt < nextt {
+				return true
+			}
+			if currt == nextt {
+				// Ignoring sample.
+				iter := heap.Pop(&c.h).(chunkenc.Iterator)
+				if iter.Next() {
+					heap.Push(&c.h, iter)
+				}
+				continue
+			}
+			heap.Push(&c.h, c.curr)
+			c.curr = heap.Pop(&c.h).(chunkenc.Iterator)
+			return true
 		}
 
-		iter := heap.Pop(&c.h).(chunkenc.Iterator)
-		if iter.Next() {
-			heap.Push(&c.h, iter)
+		if len(c.h) == 0 {
+			return false
 		}
+		c.curr = heap.Pop(&c.h).(chunkenc.Iterator)
+		return true
 	}
-
-	return len(c.h) > 0
 }
 
 func (c *chainSampleIterator) Err() error {