Reduced latency 4x for problematic case we had: 

```
BenchmarkDBBlocksQuerierSelect/[__name__="up"]
85.466431ms
99.624186ms
104.738759ms
94.795444ms
98.706406ms
96.335081ms
100.298595ms
94.806766ms
```
(was ~450ms).

Still some gap to ~60ms we had before this iterator change. I will keep looking but also I would merge this improvement straight away, to unblock release, will do next improvements in next PR.

Fixes https://github.com/prometheus/prometheus/issues/7873

Previously vs prior iterator:
```
name                                   old time/op    new time/op    delta
DBBlocksQuerierSelect/[__name__="up"]    53.0ms ± 0%   489.5ms ± 0%  +823.32%

name                                   old alloc/op   new alloc/op   delta
DBBlocksQuerierSelect/[__name__="up"]    1.86MB ± 0%    3.71MB ± 0%   +98.80%

name                                   old allocs/op  new allocs/op  delta
DBBlocksQuerierSelect/[__name__="up"]     49.5k ± 0%     67.9k ± 0%   +37.18%
```

Now:
```
name                                   old time/op    new time/op    delta
DBBlocksQuerierSelect/[__name__="up"]    52.7ms ± 0%    81.1ms ± 0%  +53.97%

name                                   old alloc/op   new alloc/op   delta
DBBlocksQuerierSelect/[__name__="up"]    1.86MB ± 0%    1.96MB ± 0%   +5.18%

name                                   old allocs/op  new allocs/op  delta
DBBlocksQuerierSelect/[__name__="up"]     49.5k ± 0%     52.4k ± 0%   +5.80%
```

Also: enhanced benchmarks.