@bwplotka 2020-09-04
What exactly is invasive? `chainSampleIterator` change?

Well in order to make sure we dont' check heap for every sample, given most of the series are not overlapping, I need to rearrange that part (: