@brian-brazil 2020-09-04
> What exactly is invasive? chainSampleIterator change?

All of it. We shouldn't need a 300 line PR to chose the right type of iterator to use.