@bwplotka 2020-09-04
That's the thing. We have one iterator not 4 like last time - that's the whole point of refactor we did.