@bwplotka 2020-09-04
Also, other lines are for benchmarking improvement, can move those to separate PR if we want.