@bwplotka 2020-09-04
Cool, will simplify.