@bwplotka 2020-09-04
Actually let's reopen, so I can preserve your comments, but will move out the critical fix.